New folder/requirements.txt
Flask>=3.0.0
pandas>=2.2.0
scikit-learn==1.6.1
joblib>=1.3.0
matplotlib>=3.8.0
openai>=1.55.0
gTTS>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
